pyproject.toml
[tool.poetry]
name = "priscillafx-website"
version = "0.1.0"
description = "Official website of Priscilla FX"
authors = ["masajinobe-ef <priscilla.effects@gmail.com>"]
readme = "README.md"
packages = [{ include = "*", from = "src" }]


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
jinja2 = "^3.1.3"
python-multipart = "^0.0.9"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
fastapi-cache2 = { extras = ["redis"], version = "^0.2.1" }
aioredis = "^2.0.1"
sqlmodel = "^0.0.16"
fastapi-users = { extras = ["sqlalchemy"], version = "^13.0.0" }
celery = "^5.4.0"
flower = "^2.0.1"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
cryptography = "^42.0.8"
loguru = "^0.7.2"


[tool.ruff]
line-length = 79
indent-width = 4


[tool.ruff.lint]
ignore = ["F403"]


[tool.ruff.format]
quote-style = "single"
indent-style = "space"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
